Move util into core
